requirements_chat.txt
gradio>=4.0.0
python-dotenv>=1.0.0
openai>=1.90.0
jinja2>=3.0.0
